Move talib file
